+++ packages/components-js/tests/vrt/puppeteer/specs/component-scaling.vrt.ts
@@ -1,7 +1,15 @@
 import { vrtTest } from '@porsche-design-system/shared/testing';
 import { getVisualRegressionStatesTester } from 'shared/src/testing/vrt';
 
-const components = ['banner', 'checkbox-wrapper', 'inline-notification', 'radio-button-wrapper'];
+const components = [
+  'banner',
+  'checkbox-wrapper',
+  'inline-notification',
+  'radio-button-wrapper',
+  'toast-basic',
+  'toast-basic-dark',
+  'toast-basic-long-text',
+];
 
 it.each(components)('should have no visual regression for scaled component %s', async (component) => {
   expect(